+++ Tests/Order.test.ps1
@@ -29,13 +29,13 @@ task Ordered {
 	$sample = "$data"
 
 	# unordered
-	$lastId = @{Value = -1}
+	$lastId = [ref]-1
 	$r = 1..100 | Split-Pipeline @param
 	"$r"
 	if ("$r" -eq $sample) { Write-Warning "Normally expected unordered data." }
 
 	# ordered
-	$lastId = @{Value = -1}
+	$lastId = [ref]-1
 	$r = 1..100 | Split-Pipeline -Order @param
 	"$r"
 	assert ("$r" -eq $sample)